// pslpwcl/src/app/modules/pslpapi/model/consultaAnnunciRequest.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CallInfo } from './callInfo';
import { OrderField } from './orderField';

export interface ConsultaAnnunciRequest { 
    callInfo: CallInfo;
    campoTestualeRicerca: string;
    flgL68Art1?: string;
    flgL68Art18?: string;
    idComune?: string;
    idCpi?: number;
    idManifestazione?: number;
    idNazioneEstera?: string;
    idServizioSpecialistico?: number;
    ordinamento?: Array<OrderField>;
    rangeKm?: number;
    tirocinio?: string;
}